Add DateTime to public
